Moved classes to their own class files
